+++ OneBlink.SDK/models/FormSubmissionEventConfiguration.cs
@@ -178,9 +178,5 @@ namespace OneBlink.SDK.Model
         {
             get; set;
         }
-        public bool isRetryable
-        {
-            get; set;
-        }
     }
 }