+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 
 ### Added
 
-- `isRetryable` to `FormSubmissionEventConfiguration` model
+- `isRetryable` to `FormSubmissionEvent` model
 
 ## [5.3.0] - 2024-01-18